Eliminate `OptionsCreate`
Move `ExportOptionsHandler`
