+++ src/export/ExportFileDialog.cpp
@@ -17,20 +17,14 @@
 #include "AllThemeResources.h"
 #include "HelpSystem.h"
 #include "AudacityMessageBox.h"
-#include "ExportOptionsEditor.h"
 #include "ExportUtils.h"
+#include "ExportOptionsHandler.h"
 
 #include <wx/bmpbuttn.h>
 #include <wx/simplebook.h>
 #include <wx/filectrl.h>
-#include <wx/checkbox.h>
-#include <wx/spinctrl.h>
-#include <wx/wupdlock.h>
 
-#include "ExportOptionsUIServices.h"
 #include "widgets/FileHistory.h"
-
-wxDEFINE_EVENT(AUDACITY_FILE_SUFFIX_EVENT, wxCommandEvent);
 
 BEGIN_EVENT_TABLE(ExportFileDialog, wxEvtHandler)
    EVT_BUTTON(wxID_HELP, ExportFileDialog::OnHelp)
@@ -53,222 +47,6 @@ bool IsExtension(const ExportPlugin& plugin, int formatIndex, const FileExtensio
 }
 
 }
-
-class ExportFileDialog::ExportOptionsHandler final
-   : public ExportOptionsEditor::Listener
-{
-public:
-
-   ExportOptionsHandler(ShuttleGui& S, ExportPlugin& plugin, int format)
-   {
-      mParent = S.GetParent();
-
-      mEditor = plugin.CreateOptionsEditor(format, this);
-      if(mEditor)
-      {
-         mEditor->Load(*gPrefs);
-         if(auto uiServices = dynamic_cast<ExportOptionsUIServices*>(mEditor.get()))
-            uiServices->PopulateUI(S);
-         else
-            PopulateOptions(S);
-      }
-      else
-         PopulateEmpty(S);
-   }
-
-   void TransferDataFromEditor()
-   {
-      if(mEditor)
-      {
-         if(auto uiServices = dynamic_cast<ExportOptionsUIServices*>(mEditor.get()))
-            uiServices->TransferDataFromWindow();
-         mEditor->Store(*gPrefs);
-      }
-   }
-
-   ExportPlugin::Parameters GetParameters() const
-   {
-      if(mEditor)
-         return ExportUtils::ParametersFromEditor(*mEditor);
-      return {};
-   }
-
-   void PopulateEmpty(ShuttleGui& S)
-   {
-      S.StartHorizontalLay(wxCENTER);
-      {
-         S.StartHorizontalLay(wxCENTER, 0);
-         {
-            S.Prop(1).AddTitle(XO("No format specific options"));
-         }
-         S.EndHorizontalLay();
-      }
-      S.EndHorizontalLay();
-   }
-
-   void PopulateOptions(ShuttleGui& S)
-   {
-      S.StartVerticalLay();
-      {
-         S.StartHorizontalLay(wxCENTER);
-         {
-            S.StartMultiColumn(2, wxCENTER);
-            {
-               for(int i = 0; i < mEditor->GetOptionsCount(); ++i)
-               {
-                  ExportOption option;
-                  if(!mEditor->GetOption(i, option))
-                     continue;
-
-                  ExportValue value;
-                  if(!mEditor->GetValue(option.id, value))
-                     continue;
-                  
-                  wxControl* control { nullptr };
-
-                  auto prompt = S.AddPrompt(option.title);
-                  if((option.flags & ExportOption::TypeMask) == ExportOption::TypeEnum)
-                  {
-                     int selected = -1;
-                     TranslatableStrings list;
-
-                     int index { 0 };
-                     std::unordered_map<int, ExportValue> indexValueMap;
-                     indexValueMap.reserve(option.values.size());
-                     for(auto& e : option.values)
-                     {
-                        list.push_back(option.names[index]);
-                        if(value == e)
-                           selected = index;
-                        indexValueMap[index] = e;
-                        ++index;
-                     }
-                     control = S.AddChoice({}, list, selected);
-                     control->Bind(wxEVT_CHOICE, [this, id = option.id, indexValueMap](const wxCommandEvent& evt)
-                     {
-                        const auto it = indexValueMap.find(evt.GetInt());
-                        if(it != indexValueMap.end())
-                           mEditor->SetValue(id, it->second);
-                     });
-                  }
-                  else if(auto selected = std::get_if<bool>(&value))
-                  {
-                     control = S.AddCheckBox({}, *selected);
-                     control->Bind(wxEVT_CHECKBOX, [this, id = option.id](const wxCommandEvent& evt)
-                     {
-                        const auto checked = evt.GetInt() != 0;
-                        mEditor->SetValue(id, checked);
-                     });
-                  }
-                  else if(auto num = std::get_if<int>(&value))
-                  {
-                     if((option.flags & ExportOption::TypeMask) == ExportOption::TypeRange)
-                     {
-                        const int min = *std::get_if<int>(&option.values[0]);
-                        const int max = *std::get_if<int>(&option.values[1]);
-                        if(max - min < 20)
-                        {
-                           control = S.AddSlider({}, *num, max, min);
-                           control->Bind(wxEVT_SLIDER, [this, id = option.id](const wxCommandEvent& evt)
-                           {
-                              mEditor->SetValue(id, evt.GetInt());
-                           });
-                        }
-                        else
-                        {
-                           control = S.AddSpinCtrl({}, *num, max, min);
-                           control->Bind(wxEVT_SPINCTRL, [this, id = option.id](const wxSpinEvent& evt)
-                           {
-                              mEditor->SetValue(id, evt.GetInt());
-                           });
-                        }
-                     }
-                     else
-                     {
-                        control = S.AddNumericTextBox({}, wxString::Format("%d", *num), 0);
-                        control->Bind(wxEVT_TEXT, [this, id = option.id](const wxCommandEvent& evt)
-                        {
-                           long num;
-                           if(evt.GetString().ToLong(&num))
-                              mEditor->SetValue(id, static_cast<int>(num));
-                        });
-                     }
-                  }
-                  else if(auto str = std::get_if<std::string>(&value))
-                  {
-                     control = S.AddTextBox({}, wxString::FromUTF8(*str), 0);
-                     control->Bind(wxEVT_TEXT, [this, id = option.id](const wxCommandEvent& evt)
-                     {
-                        mEditor->SetValue(id, evt.GetString().ToStdString());
-                     });
-                  }
-                  mIDRowIndexMap[option.id] = static_cast<int>(mRows.size());
-                  mRows.emplace_back(prompt, control);
-                  
-                  if(option.flags & ExportOption::ReadOnly)
-                     control->Disable();
-                  if(option.flags & ExportOption::Hidden)
-                  {
-                     prompt->Hide();
-                     control->Hide();
-                  }
-               }
-            }
-            S.EndMultiColumn();
-         }
-         S.EndHorizontalLay();
-      }
-      S.EndVerticalLay();
-   }
-
-   void OnExportOptionChangeBegin() override
-   {
-      mUpdateLocker = std::make_unique<wxWindowUpdateLocker>(mParent);
-   }
-
-   void OnExportOptionChangeEnd() override
-   {
-      mParent->Layout();
-      mUpdateLocker.reset();
-   }
-
-   void OnExportOptionChange(const ExportOption& option) override
-   {
-      const auto it = mIDRowIndexMap.find(option.id);
-      if(it == mIDRowIndexMap.end())
-         return;
-
-      const auto index = it->second;
-      const auto [prompt, control] = mRows[index];
-
-      const auto visible = (option.flags & ExportOption::Hidden) == 0;
-      
-      prompt->Show(visible);
-      control->Show(visible);
-      
-      const auto enabled = (option.flags & ExportOption::ReadOnly) == 0;
-      control->Enable(enabled);
-   }
-
-   void OnExtensionChange(const wxString& extension) override
-   {
-      if(mParent != nullptr)
-      {
-         // Synchronously process a change in suffix.
-         wxCommandEvent evt(AUDACITY_FILE_SUFFIX_EVENT, mParent->GetId());
-         evt.SetEventObject(mParent);
-         evt.SetString(extension);
-         mParent->ProcessWindowEvent(evt);
-      }
-   }
-
-private:
-   wxWindow* mParent { nullptr };
-   std::unique_ptr<wxWindowUpdateLocker> mUpdateLocker;
-   std::unique_ptr<ExportOptionsEditor> mEditor;
-   std::vector<std::tuple<wxStaticText*, wxControl*>> mRows;
-   std::unordered_map<int, int> mIDRowIndexMap;
-};
 
 ExportFileDialog::ExportFileDialog(wxWindow *parent,
                                    Exporter& exporter,
@@ -340,7 +118,6 @@ ExportFileDialog::ExportFileDialog(wxWindow *parent,
 }
 
 ExportFileDialog::~ExportFileDialog() = default;
-
 
 int ExportFileDialog::RunModal(wxWindow* parent,
                                Exporter& exporter,
