+++ src/export/ExportPCM.cpp
@@ -380,8 +380,7 @@ public:
 
    std::unique_ptr<ExportOptionsEditor>
    CreateOptionsEditor(int, ExportOptionsEditor::Listener*) const override;
-
-   void OptionsCreate(ShuttleGui &S, int format) override;
+   
    void Export(AudacityProject *project,
                ExportProgressListener &progressListener,
                const Parameters& parameters,
@@ -1048,10 +1047,6 @@ bool ExportPCM::AddID3Chunk(
       return false;
 #endif
    return true;
-}
-
-void ExportPCM::OptionsCreate(ShuttleGui &S, int format)
-{
 }
 
 static Exporter::RegisteredExportPlugin sRegisteredPlugin{ "PCM",
